+++ hutool-core/src/test/java/cn/hutool/core/io/FileUtilTest.java
@@ -377,4 +377,11 @@ public class FileUtilTest {
 		String mimeType = FileUtil.getMimeType("test2Write.jpg");
 		Assert.assertEquals("image/jpeg", mimeType);
 	}
+
+	@Test
+	public void isSubTest() {
+		File file = new File("d:/test");
+		File file2 = new File("d:/test2/aaa");
+		Assert.assertFalse(FileUtil.isSub(file, file2));
+	}
 }